editor css improvements
